Moved glue.qt.widgets.layer_tree_widget to glue.app.qt.widgets.layer_tree_widget, glue.qt.widgets.message_widget to glue.core.qt.message_widget, and glue.qt.widgets.settings_editor to glue.app.qt.settings_editor
