+++ glue/app/qt/application.py
@@ -17,8 +17,8 @@ from glue.qt import get_qapp
 from glue.qt.qtutil import data_wizard, load_ui, get_icon, action
 from glue.app.qt.edit_subset_mode_toolbar import EditSubsetModeToolBar
 from glue.app.qt.mdi_area import GlueMdiArea, GlueMdiSubWindow
-from glue.qt.widgets.layer_tree_widget import PlotAction, LayerTreeWidget
-from glue.qt.widgets.settings_editor import SettingsEditor
+from glue.app.qt.layer_tree_widget import PlotAction, LayerTreeWidget
+from glue.app.qt.settings_editor import SettingsEditor
 from glue.viewers.common.qt.mpl_widget import defer_draw
 from glue.viewers.common.qt.data_viewer import DataViewer
 from glue.viewers.image.qt import ImageWidget
