+++ glue/app/qt/settings_editor.py
