+++ glue/app/qt/tests/test_settings.py
