+++ glue/core/qt/data_collection_model.py
