+++ glue/core/qt/message_widget.py
