+++ glue/core/qt/message_widget.ui
