--- glue/qt/widgets/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import, division, print_function
-
-from .layer_tree_widget import LayerTreeWidget
-from .message_widget import MessageWidget
