+++ glue/viewers/common/qt/contrastlimits.ui
